1. youProductList的搜索处理 2. 铺货逻辑调整 3. 公用商品铺货测试修改

+++ mall-mbg/src/main/java/com/macro/mall/entity/XmsCustomerProduct.java
@@ -41,7 +41,7 @@ public class XmsCustomerProduct implements Serializable {
     private Long productId;
 
     @ApiModelProperty(value = "sourcing list表ID")
-    private Integer sourcingId;
+    private Long sourcingId;
 
     @ApiModelProperty(value = "货源链接")
     private String sourceLink;
